requirements.txt
alembic==0.9.10
appdirs==1.4.4
astroid==2.4.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
chromedriver-binary==91.0.4472.19.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==3.4.8
cycler==0.10.0
fake-useragent==0.1.11
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.26.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httplib2==0.19.1
hyperframe==6.0.1
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
Jinja2==3.0.1
jsonschema==3.2.0
kaitaistruct==0.9
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
lxml==4.6.3
Mako==1.1.1
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
Pillow==8.3.2
pluggy==1.0.0
proto-plus==1.19.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium-wire==4.5.2
six==1.14.0
soupsieve==2.2.1
SQLAlchemy==1.2.19
toml==0.10.0
tqdm==4.62.2
typed-ast==1.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
webdriver-manager==3.4.2
webdrivermanager==0.10.0
wrapt==1.12.1
wsproto==1.0.0
zipp==3.5.0
